+++ FelthamAlpheausA2/src/hashADT.c
@@ -1,6 +1,6 @@
 #include "hashADT.h"
 
-HashEntry * createHashEntry(int keyBase, void * element, Element * table, int tableSize, int dataSize){
+HashEntry * createHashEntry(double keyBase, void * element, Element * table, int tableSize, int dataSize){
 	HashEntry * newHash;
 	int key;
